// frontend/src/api/generated/getMeResponse.ts
/**
 * Generated by orval v7.4.1 🍺
 * Do not edit manually.
 * TeaCinema API
 * API Gateway for TeaCinema microservices
 * OpenAPI spec version: 1.0.0
 */

export interface GetMeResponse {
  /** User ID */
  id: string;
  /** User full name */
  name: string;
  /** User email address */
  email: string;
  /** User phone number */
  phone: string;
  /**
   * User avatar URL
   * @nullable
   */
  avatar: string | null;
}